requirements.txt
# Web framework
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Data validation (v1, no Rust)
pydantic<2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# AI and ML
groq>=0.9.0

# Audio processing
torch>=2.9.0
torchaudio>=2.9.0
faster-whisper>=1.0.0

# Numpy - let pip choose compatible version
numpy>=2.0.0